fix(outbox): address remaining code-review findings — 7 bugs + doc corrections
Fixes all open items from the 2026-04-18 code review. All 348 unit tests pass.

## Code fixes

**Core — publishedAny gate (Medium, #4)**
`publishedAny = true` was set unconditionally inside `catch (PartialSendException)`,
before confirming that any message actually reached the broker. A transport that always
emits PartialSendException with an empty success list would spin the publish loop at
MinPollIntervalMs. Moved the assignment inside `if (succeeded.Count > 0)`.

**Core — integer overflow in health-check thresholds (Medium, #5)**
`HeartbeatIntervalMs * 3` and `MaxPollIntervalMs * 3` are `int * int` — overflow at
~716 s intervals. Cast to `long` before multiply in OutboxHealthCheck and
OutboxPublisherOptions validation.

**SQL Server — GetPendingCount blocks under write pressure (High, #3)**
Added `WITH (NOLOCK)` to the COUNT_BIG query. GetPendingCount is a best-effort
monitoring call; dirty reads are acceptable and the table-scan should not contend with
active writers.

**SQL Server — stale-partition grace-marking gated by @ToAcquire (Medium, #6)**
The UPDATE that starts the grace timer on dead publishers' partitions was inside
`IF @ToAcquire > 0`, so publishers already at fair share never kicked off the timer.
Moved the UPDATE unconditionally before the acquire block; it now runs every rebalance
cycle regardless of whether this publisher needs more partitions. Updated README and
docs accordingly.

**SQL Server — StoreOptions snapshotted in ctor, hot-reload silently inert (Medium, #7)**
`SqlServerDbHelper` now accepts `IOptionsMonitor<SqlServerStoreOptions>` instead of a
resolved snapshot. `CommandTimeoutSeconds`, `TransientRetryMaxAttempts`, and
`TransientRetryBackoffMs` are read fresh per-call (snapshot at the start of each retry
loop for consistency within a single operation). `SqlServerOutboxStore` and
`SqlServerDeadLetterManager` updated to pass the monitor through. Tests updated to
construct the helper via the monitor helper.

## Doc corrections (D0–D6)

- **invariants.md**: removed stale "one hub per transport instance" constraint
  (EventHubOutboxTransport supports multiple hubs); corrected `RowVer` → `RowVersion`;
  clarified retry budget formula to distinguish store vs publisher MaxAttempts.
- **known-limitations.md**: removed false claim that IValidateOptions<T> is absent;
  both SqlServerStoreOptions and OutboxPublisherOptions validators already exist.
- **Core README**: corrected "Five concurrent loops" → four; fixed retry backoff
  defaults (was 500/5000 ms, actual 100/2000 ms).
- **SqlServer README**: updated Rebalance section to reflect unconditional grace-marking.
- **code-review-2026-04-18.md**: marked all findings resolved.

+++ src/Outbox.Core/Observability/OutboxHealthCheck.cs
@@ -49,7 +49,7 @@ public sealed class OutboxHealthCheck : IHealthCheck
         }
 
         // Unhealthy: heartbeat stale (3x interval = missed 2 consecutive heartbeats)
-        var heartbeatStalenessThreshold = TimeSpan.FromMilliseconds(opts.HeartbeatIntervalMs * 3);
+        var heartbeatStalenessThreshold = TimeSpan.FromMilliseconds((long)opts.HeartbeatIntervalMs * 3);
 
         if (_state.LastHeartbeatUtc != DateTimeOffset.MinValue &&
             now - _state.LastHeartbeatUtc > heartbeatStalenessThreshold)
@@ -70,7 +70,7 @@ public sealed class OutboxHealthCheck : IHealthCheck
         }
 
         // Unhealthy: no polls at all for an extended period (3x max poll interval)
-        var pollStalenessThreshold = TimeSpan.FromMilliseconds(opts.MaxPollIntervalMs * 3);
+        var pollStalenessThreshold = TimeSpan.FromMilliseconds((long)opts.MaxPollIntervalMs * 3);
 
         if (_state.LastPollUtc != DateTimeOffset.MinValue &&
             now - _state.LastPollUtc > pollStalenessThreshold)
